cloudflare_zero_trust_device_profiles & cloudflare_zero_trust_local_fallback_domain state upgraders
