+++ integration/v4_to_v5/testdata/zero_trust_device_profiles/input/zero_trust_device_profiles_e2e.tf
