--- integration/v4_to_v5/testdata/zero_trust_local_fallback_domain/input/terraform.tfstate
@@ -1,142 +1,0 @@
-{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "serial": 1,
-  "lineage": "test-lineage",
-  "outputs": {},
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_local_fallback_domain",
-      "name": "default_single",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "test-account-id",
-            "domains": [
-              {
-                "suffix": "example.com"
-              }
-            ]
-          }
-        }
-      ]
-    },
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_local_fallback_domain",
-      "name": "default_multi",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "test-account-id",
-            "domains": [
-              {
-                "suffix": "corp.example.com",
-                "description": "Corporate network",
-                "dns_server": ["10.0.0.1", "10.0.0.2"]
-              },
-              {
-                "suffix": "internal.example.com",
-                "description": "Internal services",
-                "dns_server": ["10.1.0.1"]
-              },
-              {
-                "suffix": "local.example.com"
-              }
-            ]
-          }
-        }
-      ]
-    },
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_local_fallback_domain",
-      "name": "custom_single",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "test-account-id",
-            "policy_id": "test-policy-id",
-            "domains": [
-              {
-                "suffix": "custom.example.com"
-              }
-            ]
-          }
-        }
-      ]
-    },
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_local_fallback_domain",
-      "name": "custom_multi",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "test-account-id",
-            "policy_id": "another-policy-id",
-            "domains": [
-              {
-                "suffix": "dev.example.com",
-                "description": "Development environment",
-                "dns_server": ["192.168.1.1"]
-              },
-              {
-                "suffix": "staging.example.com"
-              }
-            ]
-          }
-        }
-      ]
-    },
-    {
-      "mode": "managed",
-      "type": "cloudflare_fallback_domain",
-      "name": "deprecated_default",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "test-account-id",
-            "domains": [
-              {
-                "suffix": "deprecated.example.com",
-                "description": "Using deprecated resource name"
-              }
-            ]
-          }
-        }
-      ]
-    },
-    {
-      "mode": "managed",
-      "type": "cloudflare_fallback_domain",
-      "name": "deprecated_custom",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "test-account-id",
-            "policy_id": "deprecated-policy",
-            "domains": [
-              {
-                "suffix": "old.example.com"
-              }
-            ]
-          }
-        }
-      ]
-    }
-  ]
-}
