+++ integration/v4_to_v5/testdata/zero_trust_split_tunnel/expected/terraform.tfstate
@@ -7,6 +7,7 @@
         {
           "attributes": {
             "account_id": "test-account-123",
+            "default": true,
             "exclude": [
               {
                 "address": "192.168.0.0/16",
