--- internal/resources/zero_trust_device_default_profile/v4_to_v5_test.go
@@ -1,1827 +1,0 @@
-package zero_trust_device_default_profile
-
-import (
-	"testing"
-
-	"github.com/cloudflare/tf-migrate/internal/testhelpers"
-)
-
-func TestV4ToV5Transformation(t *testing.T) {
-	t.Run("ConfigTransformation", func(t *testing.T) {
-		migrator := NewV4ToV5Migrator()
-		tests := []testhelpers.ConfigTestCase{
-			{
-				Name: "basic default profile with minimal fields",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  name       = "Default Profile"
-  description = "Default device settings"
-  default    = true
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id                     = "f037e56e89293a057740de681ac9abbe"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "default profile with service_mode_v2 fields",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id           = "f037e56e89293a057740de681ac9abbe"
-  name                 = "Default Profile"
-  description          = "Default device settings"
-  default              = true
-  service_mode_v2_mode = "warp"
-  service_mode_v2_port = 8080
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  service_mode_v2 = {
-    mode = "warp"
-    port = 8080
-  }
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "default profile with all settings",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id              = "f037e56e89293a057740de681ac9abbe"
-  name                    = "Default Profile"
-  description             = "Default device settings"
-  default                 = true
-  allow_mode_switch       = false
-  allow_updates           = true
-  allowed_to_leave        = false
-  auto_connect            = 300
-  captive_portal          = 180
-  disable_auto_fallback   = false
-  exclude_office_ips      = true
-  support_url             = "https://support.example.com"
-  switch_locked           = false
-  tunnel_protocol         = "wireguard"
-  service_mode_v2_mode    = "warp"
-  service_mode_v2_port    = 8080
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id            = "f037e56e89293a057740de681ac9abbe"
-  allow_mode_switch     = false
-  allow_updates         = true
-  allowed_to_leave      = false
-  auto_connect          = 300
-  captive_portal        = 180
-  disable_auto_fallback = false
-  exclude_office_ips    = true
-  support_url           = "https://support.example.com"
-  switch_locked         = false
-  tunnel_protocol       = "wireguard"
-  service_mode_v2 = {
-    mode = "warp"
-    port = 8080
-  }
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-		}
-
-		testhelpers.RunConfigTransformTests(t, tests, migrator)
-	})
-
-	t.Run("ConfigTransformation_EdgeCases", func(t *testing.T) {
-		migrator := NewV4ToV5Migrator()
-		tests := []testhelpers.ConfigTestCase{
-			{
-				Name: "old resource name - cloudflare_device_settings_policy",
-				Input: `
-resource "cloudflare_device_settings_policy" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Default"
-  description = "Default policy"
-  default     = true
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id                     = "f037e56e89293a057740de681ac9abbe"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "only service_mode_v2_mode without port",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id           = "f037e56e89293a057740de681ac9abbe"
-  default              = true
-  service_mode_v2_mode = "warp"
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id                     = "f037e56e89293a057740de681ac9abbe"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "only service_mode_v2_port without mode",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id           = "f037e56e89293a057740de681ac9abbe"
-  default              = true
-  service_mode_v2_port = 8080
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  service_mode_v2 = {
-    port = 8080
-  }
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "multiple resources in one file",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "first" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  name       = "First"
-  default    = true
-}
-
-resource "cloudflare_zero_trust_device_profiles" "second" {
-  account_id  = "d138e56e89293a057740de681ac9abbf"
-  name        = "Second"
-  description = "Second profile"
-  default     = true
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "first" {
-  account_id                     = "f037e56e89293a057740de681ac9abbe"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}
-
-resource "cloudflare_zero_trust_device_default_profile" "second" {
-  account_id                     = "d138e56e89293a057740de681ac9abbf"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "resource with variables",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id = var.cloudflare_account_id
-  name       = "Default"
-  default    = true
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id                     = var.cloudflare_account_id
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "all removed fields present with various values",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  default     = true
-  name        = "Default Profile Name"
-  description = "This is a very long description that should be removed"
-  match       = "identity.email == \"user@example.com\""
-  precedence  = 100
-  enabled     = false
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id                     = "f037e56e89293a057740de681ac9abbe"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-		}
-
-		testhelpers.RunConfigTransformTests(t, tests, migrator)
-	})
-
-	t.Run("ConfigTransformation_CustomProfile", func(t *testing.T) {
-		migrator := NewV4ToV5Migrator()
-		tests := []testhelpers.ConfigTestCase{
-			{
-				Name: "custom profile with match and precedence (no default field)",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Custom Profile"
-  description = "Custom device settings"
-  match       = "identity.email == \"user@example.com\""
-  precedence  = 100
-  enabled     = true
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_custom_profile" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Custom Profile"
-  description = "Custom device settings"
-  match       = "identity.email == \"user@example.com\""
-  precedence  = 1000
-}`,
-			},
-			{
-				Name: "custom profile with default=false",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Custom Profile"
-  description = "Custom device settings"
-  default     = false
-  match       = "identity.email == \"user@example.com\""
-  precedence  = 100
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_custom_profile" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Custom Profile"
-  description = "Custom device settings"
-  match       = "identity.email == \"user@example.com\""
-  precedence  = 1000
-}`,
-			},
-			{
-				Name: "custom profile with service_mode_v2",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id           = "f037e56e89293a057740de681ac9abbe"
-  name                 = "Custom Profile"
-  match                = "identity.email == \"user@example.com\""
-  precedence           = 200
-  service_mode_v2_mode = "proxy"
-  service_mode_v2_port = 8080
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_custom_profile" "example" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  name       = "Custom Profile"
-  match      = "identity.email == \"user@example.com\""
-  precedence = 1100
-  service_mode_v2 = {
-    mode = "proxy"
-    port = 8080
-  }
-}`,
-			},
-		}
-
-		testhelpers.RunConfigTransformTests(t, tests, migrator)
-	})
-
-	t.Run("ConfigTransformation_RoutingLogic", func(t *testing.T) {
-		migrator := NewV4ToV5Migrator()
-		tests := []testhelpers.ConfigTestCase{
-			// Test all permutations of default, match, and precedence
-			// Priority: default field > presence of match+precedence
-
-			// --- default=true cases (always routes to default profile) ---
-			{
-				Name: "default=true, no match, no precedence → default profile",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Profile"
-  default     = true
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id                     = "f037e56e89293a057740de681ac9abbe"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "default=true, has match, no precedence → default profile (invalid config but default wins)",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Profile"
-  default     = true
-  match       = "identity.email == \"user@example.com\""
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id                     = "f037e56e89293a057740de681ac9abbe"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "default=true, no match, has precedence → default profile (invalid config but default wins)",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Profile"
-  default     = true
-  precedence  = 100
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id                     = "f037e56e89293a057740de681ac9abbe"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "default=true, has match, has precedence → default profile (default takes priority)",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Profile"
-  default     = true
-  match       = "identity.email == \"user@example.com\""
-  precedence  = 100
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id                     = "f037e56e89293a057740de681ac9abbe"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-
-			// --- default=false cases ---
-			{
-				Name: "default=false, no match, no precedence → default profile",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Profile"
-  default     = false
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id                     = "f037e56e89293a057740de681ac9abbe"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "default=false, has match, no precedence → default profile (missing precedence)",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Profile"
-  default     = false
-  match       = "identity.email == \"user@example.com\""
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id                     = "f037e56e89293a057740de681ac9abbe"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "default=false, no match, has precedence → default profile (missing match)",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Profile"
-  default     = false
-  precedence  = 100
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id                     = "f037e56e89293a057740de681ac9abbe"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "default=false, has match, has precedence → custom profile",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Custom Profile"
-  description = "Custom settings"
-  default     = false
-  match       = "identity.email == \"user@example.com\""
-  precedence  = 100
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_custom_profile" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Custom Profile"
-  description = "Custom settings"
-  match       = "identity.email == \"user@example.com\""
-  precedence  = 1000
-}`,
-			},
-
-			// --- no default field cases ---
-			{
-				Name: "no default, no match, no precedence → default profile",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Profile"
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id                     = "f037e56e89293a057740de681ac9abbe"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "no default, has match, no precedence → default profile (missing precedence)",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Profile"
-  match       = "identity.email == \"user@example.com\""
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id                     = "f037e56e89293a057740de681ac9abbe"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "no default, no match, has precedence → default profile (missing match)",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Profile"
-  precedence  = 100
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id                     = "f037e56e89293a057740de681ac9abbe"
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "no default, has match, has precedence → custom profile",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Custom Profile"
-  description = "Custom settings"
-  match       = "identity.email == \"user@example.com\""
-  precedence  = 150
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_custom_profile" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Custom Profile"
-  description = "Custom settings"
-  match       = "identity.email == \"user@example.com\""
-  precedence  = 1050
-}`,
-			},
-
-			// --- Edge cases with different precedence values ---
-			{
-				Name: "custom profile with precedence=1 → becomes 901",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Profile"
-  match       = "identity.email == \"test@example.com\""
-  precedence  = 1
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_custom_profile" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Profile"
-  match       = "identity.email == \"test@example.com\""
-  precedence  = 901
-}`,
-			},
-			{
-				Name: "custom profile with precedence=999 → becomes 1899",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Profile"
-  match       = "identity.email == \"test@example.com\""
-  precedence  = 999
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_custom_profile" "example" {
-  account_id  = "f037e56e89293a057740de681ac9abbe"
-  name        = "Profile"
-  match       = "identity.email == \"test@example.com\""
-  precedence  = 1899
-}`,
-			},
-		}
-
-		testhelpers.RunConfigTransformTests(t, tests, migrator)
-	})
-
-	t.Run("StateTransformation", func(t *testing.T) {
-		migrator := NewV4ToV5Migrator()
-		tests := []testhelpers.StateTestCase{
-			{
-				Name: "basic default profile state",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "example",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "name": "Default Profile",
-            "description": "Default device settings",
-            "default": true
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_default_profile",
-      "name": "example",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-			{
-				Name: "state with service_mode_v2 transformation",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "example",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "default": true,
-            "service_mode_v2_mode": "warp",
-            "service_mode_v2_port": 8080
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_default_profile",
-      "name": "example",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "service_mode_v2": {
-              "mode": "warp",
-              "port": 8080
-            }
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-			{
-				Name: "state with numeric type conversions",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "example",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "default": true,
-            "auto_connect": 300,
-            "captive_portal": 180
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_default_profile",
-      "name": "example",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "auto_connect": 300.0,
-            "captive_portal": 180.0
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-		}
-
-		testhelpers.RunStateTransformTests(t, tests, migrator)
-	})
-
-	t.Run("ConfigTransformation_SplitTunnelEmbedding", func(t *testing.T) {
-		migrator := NewV4ToV5Migrator()
-		tests := []testhelpers.ConfigTestCase{
-			{
-				Name: "default profile with single exclude split tunnel",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  name       = "Default Profile"
-  default    = true
-}
-
-resource "cloudflare_split_tunnel" "exclude_tunnel" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  mode       = "exclude"
-  tunnels {
-    address     = "192.168.1.0/24"
-    description = "Local network"
-  }
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  exclude = [{
-    address     = "192.168.1.0/24"
-    description = "Local network"
-  }]
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "default profile with single include split tunnel",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  name       = "Default Profile"
-  default    = true
-}
-
-resource "cloudflare_split_tunnel" "include_tunnel" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  mode       = "include"
-  tunnels {
-    address     = "10.0.0.0/8"
-    description = "Corporate network"
-  }
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  include = [{
-    address     = "10.0.0.0/8"
-    description = "Corporate network"
-  }]
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "default profile with multiple exclude split tunnels",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  name       = "Default Profile"
-  default    = true
-}
-
-resource "cloudflare_split_tunnel" "exclude_tunnel1" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  mode       = "exclude"
-  tunnels {
-    address     = "192.168.1.0/24"
-    description = "Network 1"
-  }
-}
-
-resource "cloudflare_split_tunnel" "exclude_tunnel2" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  mode       = "exclude"
-  tunnels {
-    address     = "10.0.0.0/8"
-    description = "Network 2"
-  }
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  exclude = [{
-    address     = "192.168.1.0/24"
-    description = "Network 1"
-  }, {
-    address     = "10.0.0.0/8"
-    description = "Network 2"
-  }]
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "default profile with both exclude and include split tunnels",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  name       = "Default Profile"
-  default    = true
-}
-
-resource "cloudflare_split_tunnel" "exclude_tunnel" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  mode       = "exclude"
-  tunnels {
-    address     = "192.168.1.0/24"
-    description = "Exclude network"
-  }
-}
-
-resource "cloudflare_split_tunnel" "include_tunnel" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  mode       = "include"
-  tunnels {
-    address     = "10.0.0.0/8"
-    description = "Include network"
-  }
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  exclude = [{
-    address     = "192.168.1.0/24"
-    description = "Exclude network"
-  }]
-  include = [{
-    address     = "10.0.0.0/8"
-    description = "Include network"
-  }]
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "default profile with split tunnels and other settings",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "example" {
-  account_id            = "f037e56e89293a057740de681ac9abbe"
-  name                  = "Default Profile"
-  default               = true
-  allow_mode_switch     = false
-  allow_updates         = true
-  tunnel_protocol       = "wireguard"
-  service_mode_v2_mode  = "warp"
-  service_mode_v2_port  = 8080
-}
-
-resource "cloudflare_split_tunnel" "exclude_tunnel" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  mode       = "exclude"
-  tunnels {
-    address     = "192.168.0.0/16"
-    description = "Private network"
-  }
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_default_profile" "example" {
-  account_id        = "f037e56e89293a057740de681ac9abbe"
-  allow_mode_switch = false
-  allow_updates     = true
-  tunnel_protocol   = "wireguard"
-  service_mode_v2 = {
-    mode = "warp"
-    port = 8080
-  }
-  exclude = [{
-    address     = "192.168.0.0/16"
-    description = "Private network"
-  }]
-  register_interface_ip_with_dns = true
-  sccm_vpn_boundary_support      = false
-}`,
-			},
-			{
-				Name: "v4 custom profile with split tunnel referencing it",
-				Input: `
-resource "cloudflare_zero_trust_device_profiles" "employees" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  name       = "Employee Profile"
-  match      = "identity.groups == \"employees\""
-  precedence = 100
-}
-
-resource "cloudflare_split_tunnel" "employee_tunnel" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  policy_id  = cloudflare_zero_trust_device_profiles.employees.id
-  mode       = "include"
-  tunnels {
-    address     = "10.100.0.0/16"
-    description = "Employee resources"
-  }
-}`,
-				Expected: `
-resource "cloudflare_zero_trust_device_custom_profile" "employees" {
-  account_id = "f037e56e89293a057740de681ac9abbe"
-  name       = "Employee Profile"
-  match      = "identity.groups == \"employees\""
-  precedence = 1000
-  include = [{
-    address     = "10.100.0.0/16"
-    description = "Employee resources"
-  }]
-}`,
-			},
-		}
-
-		testhelpers.RunConfigTransformTests(t, tests, migrator)
-	})
-
-	t.Run("StateTransformation_CustomProfile", func(t *testing.T) {
-		migrator := NewV4ToV5Migrator()
-		tests := []testhelpers.StateTestCase{
-			{
-				Name: "basic custom profile state",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "custom",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe/profile-abc-123",
-            "name": "Custom Profile",
-            "description": "Custom device settings",
-            "match": "identity.email == \"user@example.com\"",
-            "precedence": 100,
-            "enabled": true
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_custom_profile",
-      "name": "custom",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe/profile-abc-123",
-            "name": "Custom Profile",
-            "description": "Custom device settings",
-            "match": "identity.email == \"user@example.com\"",
-            "precedence": 100.0,
-            "policy_id": "profile-abc-123"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-			{
-				Name: "custom profile with policy_id already present",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "custom",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "test-policy-456",
-            "name": "Custom Profile",
-            "match": "identity.groups == \"developers\"",
-            "precedence": 200,
-            "policy_id": "test-policy-456"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_custom_profile",
-      "name": "custom",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "test-policy-456",
-            "name": "Custom Profile",
-            "match": "identity.groups == \"developers\"",
-            "precedence": 200.0,
-            "policy_id": "test-policy-456"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-			{
-				Name: "custom profile with service_mode_v2 and all fields",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "custom",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe/policy-xyz",
-            "name": "Employee Profile",
-            "description": "For employees",
-            "default": false,
-            "match": "identity.groups == \"employees\"",
-            "precedence": 150,
-            "enabled": true,
-            "auto_connect": 30,
-            "captive_portal": 180,
-            "service_mode_v2_mode": "proxy",
-            "service_mode_v2_port": 8080,
-            "tunnel_protocol": "wireguard"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_custom_profile",
-      "name": "custom",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe/policy-xyz",
-            "name": "Employee Profile",
-            "description": "For employees",
-            "match": "identity.groups == \"employees\"",
-            "precedence": 150.0,
-            "auto_connect": 30.0,
-            "captive_portal": 180.0,
-            "service_mode_v2": {
-              "mode": "proxy",
-              "port": 8080.0
-            },
-            "tunnel_protocol": "wireguard",
-            "policy_id": "policy-xyz"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-			{
-				Name: "custom profile removes only default and enabled fields",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "custom",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe/custom-123",
-            "name": "Custom Profile",
-            "description": "Should be kept",
-            "default": false,
-            "match": "identity.email == \"test@example.com\"",
-            "precedence": 50,
-            "enabled": false
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_custom_profile",
-      "name": "custom",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe/custom-123",
-            "name": "Custom Profile",
-            "description": "Should be kept",
-            "match": "identity.email == \"test@example.com\"",
-            "precedence": 50.0,
-            "policy_id": "custom-123"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-		}
-
-		testhelpers.RunStateTransformTests(t, tests, migrator)
-	})
-
-	t.Run("StateTransformation_MultipleResources", func(t *testing.T) {
-		migrator := NewV4ToV5Migrator()
-		tests := []testhelpers.StateTestCase{
-			{
-				Name: "multiple resources - mix of default and custom profiles",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "default",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe",
-            "name": "Default Profile",
-            "default": true,
-            "auto_connect": 15
-          }
-        }
-      ]
-    },
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "custom_one",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe/policy-one",
-            "name": "Custom One",
-            "match": "identity.groups == \"group1\"",
-            "precedence": 100,
-            "captive_portal": 180
-          }
-        }
-      ]
-    },
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "custom_two",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe/policy-two",
-            "name": "Custom Two",
-            "match": "identity.groups == \"group2\"",
-            "precedence": 200,
-            "allow_mode_switch": false
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_default_profile",
-      "name": "default",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe",
-            "auto_connect": 15.0
-          }
-        }
-      ]
-    },
-    {
-      "type": "cloudflare_zero_trust_device_custom_profile",
-      "name": "custom_one",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe/policy-one",
-            "name": "Custom One",
-            "match": "identity.groups == \"group1\"",
-            "precedence": 100.0,
-            "captive_portal": 180.0,
-            "policy_id": "policy-one"
-          }
-        }
-      ]
-    },
-    {
-      "type": "cloudflare_zero_trust_device_custom_profile",
-      "name": "custom_two",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe/policy-two",
-            "name": "Custom Two",
-            "match": "identity.groups == \"group2\"",
-            "precedence": 200.0,
-            "allow_mode_switch": false,
-            "policy_id": "policy-two"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-			{
-				Name: "multiple resources - all default profiles",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "first",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "account-111",
-            "id": "profile-111",
-            "name": "First",
-            "default": true
-          }
-        }
-      ]
-    },
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "second",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "account-222",
-            "id": "profile-222",
-            "name": "Second",
-            "default": true
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_default_profile",
-      "name": "first",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "account-111",
-            "id": "profile-111"
-          }
-        }
-      ]
-    },
-    {
-      "type": "cloudflare_zero_trust_device_default_profile",
-      "name": "second",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "account-222",
-            "id": "profile-222"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-			{
-				Name: "multiple resources - all custom profiles",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "employees",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe/emp-policy",
-            "name": "Employees",
-            "match": "identity.groups == \"employees\"",
-            "precedence": 100
-          }
-        }
-      ]
-    },
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "contractors",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe/con-policy",
-            "name": "Contractors",
-            "match": "identity.groups == \"contractors\"",
-            "precedence": 200
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_custom_profile",
-      "name": "employees",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe/emp-policy",
-            "name": "Employees",
-            "match": "identity.groups == \"employees\"",
-            "precedence": 100.0,
-            "policy_id": "emp-policy"
-          }
-        }
-      ]
-    },
-    {
-      "type": "cloudflare_zero_trust_device_custom_profile",
-      "name": "contractors",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f037e56e89293a057740de681ac9abbe/con-policy",
-            "name": "Contractors",
-            "match": "identity.groups == \"contractors\"",
-            "precedence": 200.0,
-            "policy_id": "con-policy"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-		}
-
-		testhelpers.RunStateTransformTests(t, tests, migrator)
-	})
-
-	t.Run("StateTransformation_EdgeCases", func(t *testing.T) {
-		migrator := NewV4ToV5Migrator()
-		tests := []testhelpers.StateTestCase{
-			{
-				Name: "old resource name - cloudflare_device_settings_policy",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_device_settings_policy",
-      "name": "example",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "default": true
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_default_profile",
-      "name": "example",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-			{
-				Name: "only service_mode_v2_mode field",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "example",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "service_mode_v2_mode": "warp"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_default_profile",
-      "name": "example",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-			{
-				Name: "only service_mode_v2_port field",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "example",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "service_mode_v2_port": 9000
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_default_profile",
-      "name": "example",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "service_mode_v2": {
-              "port": 9000
-            }
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-			{
-				Name: "all fields with type conversions",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "example",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "default": true,
-            "name": "Default Profile",
-            "description": "Default settings",
-            "match": "identity.email == \"test@example.com\"",
-            "precedence": 10,
-            "enabled": true,
-            "auto_connect": 0,
-            "captive_portal": 300,
-            "allow_mode_switch": true,
-            "allow_updates": false,
-            "allowed_to_leave": true,
-            "disable_auto_fallback": false,
-            "exclude_office_ips": true,
-            "support_url": "https://support.example.com",
-            "switch_locked": false,
-            "tunnel_protocol": "wireguard",
-            "service_mode_v2_mode": "warp",
-            "service_mode_v2_port": 8080
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_default_profile",
-      "name": "example",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "auto_connect": 0,
-            "captive_portal": 300,
-            "allow_mode_switch": true,
-            "allow_updates": false,
-            "allowed_to_leave": true,
-            "disable_auto_fallback": false,
-            "exclude_office_ips": true,
-            "support_url": "https://support.example.com",
-            "switch_locked": false,
-            "tunnel_protocol": "wireguard",
-            "service_mode_v2": {
-              "mode": "warp",
-              "port": 8080
-            }
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-			{
-				Name: "minimal state - no optional fields",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "example",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "default": true
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_default_profile",
-      "name": "example",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-			{
-				Name: "state with zero values for numeric fields",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "example",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "auto_connect": 0,
-            "captive_portal": 0
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_default_profile",
-      "name": "example",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "auto_connect": 0,
-            "captive_portal": 0
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-			{
-				Name: "empty instance attributes",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "example",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {}
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_default_profile",
-      "name": "example",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {}
-        }
-      ]
-    }
-  ]
-}`,
-			},
-			{
-				Name: "state with fallback_domains and empty exclude array - both should be removed",
-				Input: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "cloudflare_zero_trust_device_profiles",
-      "name": "example",
-      "provider": "provider[\"registry.terraform.io/cloudflare/cloudflare\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "default": true,
-            "fallback_domains": [
-              {
-                "suffix": "corp.example.com",
-                "description": "Corporate network",
-                "dns_server": ["10.0.0.1", "10.0.0.2"]
-              },
-              {
-                "suffix": "internal.example.com",
-                "description": "Internal services",
-                "dns_server": ["10.1.0.1"]
-              }
-            ],
-            "exclude": [],
-            "tunnel_protocol": "wireguard"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-				Expected: `{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "resources": [
-    {
-      "type": "cloudflare_zero_trust_device_default_profile",
-      "name": "example",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "account_id": "f037e56e89293a057740de681ac9abbe",
-            "id": "f174e90a-fafe-4643-bbbc-4a0ed4fc8415",
-            "tunnel_protocol": "wireguard"
-          }
-        }
-      ]
-    }
-  ]
-}`,
-			},
-		}
-
-		testhelpers.RunStateTransformTests(t, tests, migrator)
-	})
-}
