+++ internal/resources/zero_trust_local_fallback_domain/v4_to_v5.go
@@ -6,7 +6,6 @@ import (
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclwrite"
 	"github.com/tidwall/gjson"
-	"github.com/tidwall/sjson"
 
 	"github.com/cloudflare/tf-migrate/internal"
 	"github.com/cloudflare/tf-migrate/internal/transform"
@@ -142,6 +141,7 @@ func (m *V4ToV5Migrator) GetResourceRename() (string, string) {
 }
 
 func (m *V4ToV5Migrator) TransformConfig(ctx *transform.Context, block *hclwrite.Block) (*transform.TransformResult, error) {
+	resourceName := tfhcl.GetResourceName(block)
 	body := block.Body()
 
 	// Check if policy_id exists and is not null
@@ -166,6 +166,7 @@ func (m *V4ToV5Migrator) TransformConfig(ctx *transform.Context, block *hclwrite
 
 	// Rename resource type to appropriate v5 resource
 	currentType := tfhcl.GetResourceType(block)
+	oldType := currentType // Store original type for moved block
 	tfhcl.RenameResourceType(block, currentType, newResourceType)
 
 	// Remove policy_id attribute if it's null (default profile doesn't accept policy_id)
@@ -176,77 +177,25 @@ func (m *V4ToV5Migrator) TransformConfig(ctx *transform.Context, block *hclwrite
 	// Convert domains blocks to attribute array
 	tfhcl.ConvertBlocksToAttributeList(body, "domains", nil)
 
+	// Generate moved block
+	from := oldType + "." + resourceName
+	to := newResourceType + "." + resourceName
+	movedBlock := tfhcl.CreateMovedBlock(from, to)
+
 	return &transform.TransformResult{
-		Blocks:         []*hclwrite.Block{block},
-		RemoveOriginal: false,
+		Blocks:         []*hclwrite.Block{block, movedBlock},
+		RemoveOriginal: true,
 	}, nil
 }
 
+// UsesProviderStateUpgrader indicates that this resource uses provider-based state migration
+func (m *V4ToV5Migrator) UsesProviderStateUpgrader() bool {
+	return true
+}
+
 func (m *V4ToV5Migrator) TransformState(ctx *transform.Context, stateJSON gjson.Result, resourcePath, resourceName string) (string, error) {
-	result := stateJSON.String()
-
-	attrs := stateJSON.Get("attributes")
-	if !attrs.Exists() {
-		result, _ = sjson.Set(result, "schema_version", 0)
-		m.lastTransformedType = m.newTypeDefault
-		return result, nil
-	}
-
-	// Check if policy_id exists and is not null
-	policyID := attrs.Get("policy_id")
-	hasPolicyID := policyID.Exists() && policyID.Type != gjson.Null
-	if hasPolicyID {
-		m.lastTransformedType = m.newTypeCustom
-	} else {
-		m.lastTransformedType = m.newTypeDefault
-	}
-
-	transform.SetStateTypeRename(ctx, resourceName, m.oldType, m.lastTransformedType)
-	transform.SetStateTypeRename(ctx, resourceName, m.oldTypeDeprecated, m.lastTransformedType)
-
-	// Remove policy_id from state if it's null (default profile doesn't have policy_id)
-	if !hasPolicyID && policyID.Exists() {
-		result, _ = sjson.Delete(result, "attributes.policy_id")
-	}
-
-	// Transform ID format for custom profile
-	// v4 ID format: account_id/policy_id
-	// v5 ID format: policy_id (just the policy_id portion)
-	if hasPolicyID {
-		if id := attrs.Get("id"); id.Exists() {
-			idStr := id.String()
-			// Split on "/" to get the policy_id portion
-			if slashIdx := strings.Index(idStr, "/"); slashIdx != -1 && slashIdx < len(idStr)-1 {
-				newID := idStr[slashIdx+1:]
-				result, _ = sjson.Set(result, "attributes.id", newID)
-			}
-		}
-	}
-
-	// Transform domains (TypeSet → ListNestedAttribute/SetNestedAttribute)
-	// The state structure is the same for both v5 variants, just different collection semantics
-	// v4: domains is a set of objects
-	// v5: domains is either a list (default profile) or set (custom profile) of objects
-	// The JSON structure is identical, so no transformation needed
-	domains := attrs.Get("domains")
-	if domains.Exists() && domains.IsArray() {
-		if len(domains.Array()) == 0 {
-			result, _ = sjson.Delete(result, "attributes.domains")
-		}
-	}
-
-	// Handle dns_server empty arrays within each domain
-	if domains.Exists() && domains.IsArray() {
-		domains.ForEach(func(key, value gjson.Result) bool {
-			dnsServer := value.Get("dns_server")
-			if dnsServer.Exists() && dnsServer.IsArray() && len(dnsServer.Array()) == 0 {
-				result, _ = sjson.Delete(result, "attributes.domains."+key.String()+".dns_server")
-			}
-			return true
-		})
-	}
-
-	result, _ = sjson.Set(result, "schema_version", 0)
-
-	return result, nil
+	// State transformation is handled by the provider's StateUpgraders (MoveState/UpgradeState)
+	// The moved block generated in TransformConfig triggers the provider's migration logic
+	// This function is a no-op for zero_trust_local_fallback_domain migration
+	return stateJSON.String(), nil
 }
