+++ internal/resources/zero_trust_local_fallback_domain/v4_to_v5_test.go
@@ -29,6 +29,11 @@ resource "cloudflare_zero_trust_device_default_profile_local_domain_fallback" "e
       suffix = "example.com"
     }
   ]
+}
+
+moved {
+  from = cloudflare_zero_trust_local_fallback_domain.example
+  to   = cloudflare_zero_trust_device_default_profile_local_domain_fallback.example
 }`,
 		},
 		{
@@ -54,6 +59,11 @@ resource "cloudflare_zero_trust_device_default_profile_local_domain_fallback" "e
       dns_server  = ["1.1.1.1", "8.8.8.8"]
     }
   ]
+}
+
+moved {
+  from = cloudflare_zero_trust_local_fallback_domain.example
+  to   = cloudflare_zero_trust_device_default_profile_local_domain_fallback.example
 }`,
 		},
 		{
@@ -93,6 +103,11 @@ resource "cloudflare_zero_trust_device_default_profile_local_domain_fallback" "e
       suffix = "third.com"
     }
   ]
+}
+
+moved {
+  from = cloudflare_zero_trust_local_fallback_domain.example
+  to   = cloudflare_zero_trust_device_default_profile_local_domain_fallback.example
 }`,
 		},
 		{
@@ -114,6 +129,11 @@ resource "cloudflare_zero_trust_device_default_profile_local_domain_fallback" "e
       suffix = "example.com"
     }
   ]
+}
+
+moved {
+  from = cloudflare_fallback_domain.example
+  to   = cloudflare_zero_trust_device_default_profile_local_domain_fallback.example
 }`,
 		},
 		{
@@ -136,6 +156,11 @@ resource "cloudflare_zero_trust_device_default_profile_local_domain_fallback" "e
       suffix = "example.com"
     }
   ]
+}
+
+moved {
+  from = cloudflare_zero_trust_local_fallback_domain.example
+  to   = cloudflare_zero_trust_device_default_profile_local_domain_fallback.example
 }`,
 		},
 		{
@@ -159,6 +184,11 @@ resource "cloudflare_zero_trust_device_custom_profile_local_domain_fallback" "ex
       suffix = "example.com"
     }
   ]
+}
+
+moved {
+  from = cloudflare_zero_trust_local_fallback_domain.example
+  to   = cloudflare_zero_trust_device_custom_profile_local_domain_fallback.example
 }`,
 		},
 		{
@@ -186,6 +216,11 @@ resource "cloudflare_zero_trust_device_custom_profile_local_domain_fallback" "ex
       dns_server  = ["10.0.0.1"]
     }
   ]
+}
+
+moved {
+  from = cloudflare_zero_trust_local_fallback_domain.example
+  to   = cloudflare_zero_trust_device_custom_profile_local_domain_fallback.example
 }`,
 		},
 		{
@@ -220,6 +255,11 @@ resource "cloudflare_zero_trust_device_custom_profile_local_domain_fallback" "ex
       dns_server  = ["10.0.0.2", "10.0.0.3"]
     }
   ]
+}
+
+moved {
+  from = cloudflare_zero_trust_local_fallback_domain.example
+  to   = cloudflare_zero_trust_device_custom_profile_local_domain_fallback.example
 }`,
 		},
 		{
@@ -243,6 +283,11 @@ resource "cloudflare_zero_trust_device_custom_profile_local_domain_fallback" "ex
       suffix = "example.com"
     }
   ]
+}
+
+moved {
+  from = cloudflare_fallback_domain.example
+  to   = cloudflare_zero_trust_device_custom_profile_local_domain_fallback.example
 }`,
 		},
 		{
@@ -275,6 +320,11 @@ resource "cloudflare_zero_trust_device_default_profile_local_domain_fallback" "d
   ]
 }
 
+moved {
+  from = cloudflare_zero_trust_local_fallback_domain.default
+  to   = cloudflare_zero_trust_device_default_profile_local_domain_fallback.default
+}
+
 resource "cloudflare_zero_trust_device_custom_profile_local_domain_fallback" "custom" {
   account_id = "f037e56e89293a057740de681ac9abbe"
   policy_id  = "policy123"
@@ -284,302 +334,16 @@ resource "cloudflare_zero_trust_device_custom_profile_local_domain_fallback" "cu
       suffix = "custom.example.com"
     }
   ]
+}
+
+moved {
+  from = cloudflare_zero_trust_local_fallback_domain.custom
+  to   = cloudflare_zero_trust_device_custom_profile_local_domain_fallback.custom
 }`,
 		},
 	}
 
 	testhelpers.RunConfigTransformTests(t, tests, migrator)
-}
-
-func TestStateTransformation(t *testing.T) {
-	migrator := NewV4ToV5Migrator()
-
-	tests := []testhelpers.StateTestCase{
-		{
-			Name: "basic default profile state",
-			Input: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "domains": [
-      {
-        "suffix": "example.com"
-      }
-    ]
-  }
-}`,
-			Expected: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "domains": [
-      {
-        "suffix": "example.com"
-      }
-    ]
-  }
-}`,
-		},
-		{
-			Name: "default profile with all fields",
-			Input: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "domains": [
-      {
-        "suffix": "example.com",
-        "description": "Example domain",
-        "dns_server": ["1.1.1.1", "8.8.8.8"]
-      }
-    ]
-  }
-}`,
-			Expected: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "domains": [
-      {
-        "suffix": "example.com",
-        "description": "Example domain",
-        "dns_server": ["1.1.1.1", "8.8.8.8"]
-      }
-    ]
-  }
-}`,
-		},
-		{
-			Name: "default profile with multiple domains",
-			Input: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "domains": [
-      {
-        "suffix": "example.com",
-        "description": "Example"
-      },
-      {
-        "suffix": "another.com",
-        "dns_server": ["1.1.1.1"]
-      },
-      {
-        "suffix": "third.com"
-      }
-    ]
-  }
-}`,
-			Expected: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "domains": [
-      {
-        "suffix": "example.com",
-        "description": "Example"
-      },
-      {
-        "suffix": "another.com",
-        "dns_server": ["1.1.1.1"]
-      },
-      {
-        "suffix": "third.com"
-      }
-    ]
-  }
-}`,
-		},
-		{
-			Name: "default profile with null policy_id",
-			Input: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "policy_id": null,
-    "domains": [
-      {
-        "suffix": "example.com"
-      }
-    ]
-  }
-}`,
-			Expected: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "domains": [
-      {
-        "suffix": "example.com"
-      }
-    ]
-  }
-}`,
-		},
-		{
-			Name: "basic custom profile state",
-			Input: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "policy_id": "policy123",
-    "domains": [
-      {
-        "suffix": "corp.example.com"
-      }
-    ]
-  }
-}`,
-			Expected: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "policy_id": "policy123",
-    "domains": [
-      {
-        "suffix": "corp.example.com"
-      }
-    ]
-  }
-}`,
-		},
-		{
-			Name: "custom profile with all fields",
-			Input: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "policy_id": "policy123",
-    "domains": [
-      {
-        "suffix": "corp.example.com",
-        "description": "Corporate domain",
-        "dns_server": ["10.0.0.1"]
-      }
-    ]
-  }
-}`,
-			Expected: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "policy_id": "policy123",
-    "domains": [
-      {
-        "suffix": "corp.example.com",
-        "description": "Corporate domain",
-        "dns_server": ["10.0.0.1"]
-      }
-    ]
-  }
-}`,
-		},
-		{
-			Name: "empty domains array",
-			Input: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "domains": []
-  }
-}`,
-			Expected: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe"
-  }
-}`,
-		},
-		{
-			Name: "empty dns_server in domain",
-			Input: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "domains": [
-      {
-        "suffix": "example.com",
-        "dns_server": []
-      }
-    ]
-  }
-}`,
-			Expected: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "domains": [
-      {
-        "suffix": "example.com"
-      }
-    ]
-  }
-}`,
-		},
-		{
-			Name: "missing optional fields",
-			Input: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "domains": [
-      {
-        "suffix": "example.com"
-      }
-    ]
-  }
-}`,
-			Expected: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "domains": [
-      {
-        "suffix": "example.com"
-      }
-    ]
-  }
-}`,
-		},
-		{
-			Name: "invalid instance - missing attributes",
-			Input: `{
-  "schema_version": 0
-}`,
-			Expected: `{
-  "schema_version": 0
-}`,
-		},
-		{
-			Name: "null description field",
-			Input: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "domains": [
-      {
-        "suffix": "example.com",
-        "description": null
-      }
-    ]
-  }
-}`,
-			Expected: `{
-  "schema_version": 0,
-  "attributes": {
-    "account_id": "f037e56e89293a057740de681ac9abbe",
-    "domains": [
-      {
-        "suffix": "example.com",
-        "description": null
-      }
-    ]
-  }
-}`,
-		},
-	}
-
-	testhelpers.RunStateTransformTests(t, tests, migrator)
 }
 
 func TestCanHandle(t *testing.T) {
